Time: 15 ms (92.19%), Space: 22.5 MB (72.83%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -20,7 +20,7 @@ public:
         while(!q.empty()) {
             int size = q.size();
             int i = size;
-            double sum = 0;
+            long  sum = 0;
             while(i--) {
                 TreeNode * tmp = q.front();
                 q.pop();
